+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased][]
 
+## [2.2.0][] - 2019-10-11
+
 ### Added
 - GraphQL schema for wvw_score feature
 - GraphQL schema for bot guild
@@ -40,7 +42,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Store keys in sqlite instead of redis
 
 
-[Unreleased]: https://github.com/Nabrok/gw2-discord-bot/compare/v2.1.2...HEAD
+[Unreleased]: https://github.com/Nabrok/gw2-discord-bot/compare/v2.2.0...HEAD
+[2.2.0]: https://github.com/Nabrok/gw2-discord-bot/compare/v2.1.2...v2.2.0
 [2.1.2]: https://github.com/Nabrok/gw2-discord-bot/compare/v2.1.1...v2.1.2
 [2.1.1]: https://github.com/Nabrok/gw2-discord-bot/compare/v2.1.0...v2.1.1
 [2.1.0]: https://github.com/Nabrok/gw2-discord-bot/compare/v2.0.0...v2.1.0